--- RCHIWebDash/RCHIWebDashBoard/src/components/CityTables/constants.js
@@ -1,297 +1,0 @@
-export const citynames = {
-
-    2019 :[
-    'BANNING',				
-    'BEAUMONT',
-    'BLYTHE',
-    'CALIMESA',
-    'CATHEDRAL CITY',
-    'COACHELLA',
-    'CORONA',
-    'DESERT HOT SPRINGS',
-    'HEMET',
-    'INDIAN WELLS',
-    'INDIO',
-    'JURUPA VALLEY',
-    'LA QUINTA',
-    'LAKE ELSINORE',
-    'MENIFEE',
-    'MORENO VALLEY',
-    'MURRIETA',
-    'NORCO',
-    'PALM DESERT',
-    'PALM SPRINGS',
-    'PERRIS',
-    'RANCHO MIRAGE',
-    'RIVERSIDE',
-    'RIVERSIDE 2',
-    'RIVERSIDE 1',
-    'SAN JACINTO',
-    'TEMECULA',
-    'WILDOMAR',
-    'UNINCORPORATED 1',
-    'UNINCORPORATED 2',
-    'UNINCORPORATED 3',
-    'UNINCORPORATED 4',
-    'UNINCORPORATED 5',
-    ],
-
-    2020 : [
-    'RIVERSIDE 1',
-    'LAKE ELSINORE',
-    'UNINCORPORATED DISTRICT 1',
-    'WILDOMAR',
-    'CANYON LAKE',
-    'NORCO',
-    'CORONA',
-    'RIVERSIDE 2',
-    'JURUPA VALLEY',
-    'UNINCORPORATED DISTRICT 2',
-    'EASTVALE',
-    'TEMECULA',
-    'UNINCORPORATED DISTRICT 3',
-    'MURRIETA',
-    'HEMET',
-    'SAN JACINTO',
-    'DESERT HOT SPRINGS',
-    'UNINCORPORATED DISTRICT 4',
-    'LA QUINTA',
-    'COACHELLA',
-    'INDIAN WELLS',
-    'INDIO',
-    'PALM SPRINGS',
-    'CATHEDRAL CITY',
-    'BLYTHE',
-    'PALM DESERT',
-    'RANCHO MIRAGE',
-    'MECCA',
-    'MENIFEE',
-    'PERRIS',
-    'UNINCORPORATED DISTRICT 5',
-    'BEAUMONT',
-    'MORENO VALLEY',
-    'CALIMESA',
-    'BANNING',
-    'RIVERSIDE',
-    ]
-};
-
-export const filterData = [
-    "Unknown Veteran" 
-    , "Adults Only", 
-    'Children Only', 
-    'Families with Children',
-    'Pet Owners', 'Individuals',
-    'Newly Homeless',
-    'Seniors 60+',
-    'Woods',
-    'Encampment', 
-    'Couch', 
-    'Vehicle',
-    'UnderBridge', 
-    'Street', 
-    'Park', 
-    'Other',
-    'Bus',
-    'Abandoned Building'
-]
-
-export const base = {
-    'District': 0,
-    'Age':	{
-        'Youth (18-24)': { 'Interview':0, 'Observation':0 },
-        'Children (<18)': { 'Interview':0, 'Observation':0 },
-        'Adults (>24)': { 'Interview':0, 'Observation':0 },
-        //'Elderly (62+)': { 'Interview':0, 'Observation':0 },
-        'Unknown Age': { 'Interview':0, 'Observation':0 },
-        'Total': { 'Interview':0, 'Observation':0 },
-    },
-    'Race':	{
-        'Asian': { 'Interview':0, 'Observation':0 },
-        'Black': { 'Interview':0, 'Observation':0 },
-        'White': { 'Interview':0, 'Observation':0 },
-        'American Indian': { 'Interview':0, 'Observation':0 },
-        'Native Hawaiian': { 'Interview':0, 'Observation':0 },
-        'Multiple Races': { 'Interview':0, 'Observation':0 },
-        'Unknown Race': { 'Interview':0, 'Observation':0 },
-        'Total': { 'Interview':0, 'Observation':0 },
-    },
-    'Ethnicity':	{
-        'Hispanic': {'Interview':0, 'Observation':0 },
-        'NonHispanic': { 'Interview':0, 'Observation':0 },
-        'Unknown Ethnicity': { 'Interview':0, 'Observation':0 },
-        'Total': { 'Interview':0, 'Observation':0 },
-    },
-    'Gender':	{
-        'Male': { 'Interview':0, 'Observation':0 },
-        'Female': { 'Interview':0, 'Observation':0 },
-        'Transgender': { 'Interview':0, 'Observation':0 },
-        'Gender Non-Conforming': { 'Interview':0, 'Observation':0 },
-        'Unknown Gender': { 'Interview':0, 'Observation':0 },
-        'Total': { 'Interview':0, 'Observation':0 },
-    },
-    'Subpopulations':	{
-        'Veteran Yes': { 'Interview':0, 'Observation':0 },
-        'Veteran No': { 'Interview':0, 'Observation':0 },
-        'Chronically Homeless': { 'Interview':0, 'Observation':0 },
-        'Not Chronically Homeless': { 'Interview':0, 'Observation':0 },
-        'Substance Abuse': { 'Interview':0, 'Observation':0 },
-        'No Substance Abuse': { 'Interview':0, 'Observation':0 },
-        'Unknown Substance Abuse': { 'Interview':0, 'Observation':0 },
-        'PTSD': { 'Interview':0, 'Observation':0 },
-        'No PTSD': { 'Interview':0, 'Observation':0 },
-        'Unknown PTSD': { 'Interview':0, 'Observation':0 },
-        'Mental Health Conditions': { 'Interview':0, 'Observation':0 },
-        'No Mental Health Conditions': { 'Interview':0, 'Observation':0 },
-        'Unknown Mental Health Conditions': { 'Interview':0, 'Observation':0 },
-        'Physical Disability': { 'Interview':0, 'Observation':0 },
-        'No Physical Disability': { 'Interview':0, 'Observation':0 },
-        'Unknown Physical Disability': { 'Interview':0, 'Observation':0 },
-        'Developmental Disability': { 'Interview':0, 'Observation':0 },
-        'No Developmental Disability': { 'Interview':0, 'Observation':0 },
-        'Unknown Developmental Disability': { 'Interview':0, 'Observation':0 },
-        'Brain Injury': { 'Interview':0, 'Observation':0 },
-        'No Brain Injury': { 'Interview':0, 'Observation':0 },
-        'Unknown Brain Injury': { 'Interview':0, 'Observation':0 },
-        'Victim of Domestic Violence': { 'Interview':0, 'Observation':0 },
-        'Not Victim of Domestic Violence': { 'Interview':0, 'Observation':0 },
-        'Unknown Victim of Domestic Violence': { 'Interview':0, 'Observation':0 },
-        'AIDS or HIV': { 'Interview':0, 'Observation':0 },
-        'No AIDS or HIV': { 'Interview':0, 'Observation':0 },
-        'Unknown AIDS or HIV': { 'Interview':0, 'Observation':0 },
-        'Jail Release 90 Days: Probation': { 'Interview':0, 'Observation':0 },
-        'Jail Release 90 Days: Parole': { 'Interview':0, 'Observation':0 },
-        'Jail Release 90 Days: Completed Sentence': { 'Interview':0, 'Observation':0 },
-        'Jail Release 90 Days: (Unspecified)': { 'Interview':0, 'Observation':0 },
-        'Jail Release 12 Months: Probation': { 'Interview':0, 'Observation':0 },
-        'Jail Release 12 Months: Parole': { 'Interview':0, 'Observation':0 },
-        'Jail Release 12 Months: Completed Sentence': { 'Interview':0, 'Observation':0 },
-        'Jail Release 12 Months: (Unspecified)': { 'Interview':0, 'Observation':0 },
-        'No Jail': { 'Interview':0, 'Observation':0 },
-        'Unknown Jail': { 'Interview':0, 'Observation':0 },
-        //'Families with Children': { 'Interview':0, 'Observation':0 },
-        'Total': { 'Interview':0, 'Observation':0 },
-    }
-};
-
-export const SubPopHeaders = [
-    'Veterans',
-    'Chronically Homeless',
-    'Substance Abuse',
-    'PTSD',
-    'Mental Health Conditions',
-    'Physical Disability',
-    'Developmental Disability',
-    'Brain Injury',
-    'Victim of Domestic Violence',
-    'AIDS or HIV',
-    'Jail',
-    //'Families with Children'
-]
-
-export const Titles = {
-    'Race': [ //7
-        'American Indian',
-        'Asian',
-        'Black',
-        'Native Hawaiian',
-        'White',
-        'Multiple Races',
-        'Unknown Race',
-        'Total', 
-    ],
-    'Ethnicity': [ //4
-        'Hispanic',
-        'NonHispanic',
-        'Unknown Ethnicity',
-        'Total',
-    ],
-    'Gender': [ //6
-        'Male',
-        'Female',
-        'Transgender',
-        'Gender Non-Conforming',
-        'Unknown Gender',
-        'Total',
-    ],
-    'Age': [ //5
-        'Adults (>24)',
-        'Youth (18-24)',
-        'Children (<18)',
-        //'Elderly (62+)',
-        'Unknown Age',
-        'Total',
-    ],
-    
-    'Veterans': [
-        'Veteran Yes',
-        'Veteran No',
-    ],
-    'Chronically Homeless': [
-        'Chronically Homeless',
-        'Not Chronically Homeless',
-    ],
-    'Substance Abuse': [
-        'Substance Abuse',
-        'No Substance Abuse',
-        'Unknown Substance Abuse',
-    ],
-    'PTSD': [
-        'PTSD',
-        'No PTSD',
-        'Unknown PTSD',
-    ],
-    'Mental Health Conditions': [
-        'Mental Health Conditions',
-        'No Mental Health Conditions',
-        'Unknown Mental Health Conditions',
-    ],
-    'Physical Disability': [
-        'Physical Disability',
-        'No Physical Disability',
-        'Unknown Physical Disability',
-    ],
-    'Developmental Disability': [
-        'Developmental Disability',
-        'No Developmental Disability',
-        'Unknown Developmental Disability',
-    ],
-    'Brain Injury': [
-        'Brain Injury',
-        'No Brain Injury',
-        'Unknown Brain Injury',
-    ],
-    'Victim of Domestic Violence': [
-        'Victim of Domestic Violence',
-        'Not Victim of Domestic Violence',
-        'Unknown Victim of Domestic Violence',
-    ],
-    'AIDS or HIV': [
-        'AIDS or HIV',
-        'No AIDS or HIV',
-        'Unknown AIDS or HIV',
-    ],
-    'Jail': [
-        'Jail Release 90 Days: Probation',
-        'Jail Release 90 Days: Parole',
-        'Jail Release 90 Days: Completed Sentence',
-        'Jail Release 90 Days: (Unspecified)',
-        'Jail Release 12 Months: Probation',
-        'Jail Release 12 Months: Parole',
-        'Jail Release 12 Months: Completed Sentence',
-        'Jail Release 12 Months: (Unspecified)',
-        'No Jail',
-        'Unknown Jail',
-    ],
-    // 'Families with Children' : [
-    //     'Families with Children',
-    // ]
-}
-
-export const Catergories = [
-    'Race',
-    'Ethnicity',
-    'Gender',
-    'Age',
-    'Subpopulations'
-]
